+++ src/mqtt-plus-source.ts
@@ -33,7 +33,7 @@ import { nanoid }                                         from "nanoid"
 /*  internal requirements  */
 import { CreditGate, ReadableTee,
     sendBufferAsChunks, sendStreamAsChunks,
-    makeMutuallyExclusiveFields }                         from "./mqtt-plus-util"
+    makeMutuallyExclusiveFields, alignHandlerParams }     from "./mqtt-plus-util"
 import { run, Spool, ensureError }                        from "./mqtt-plus-error"
 import type { SourceFetchRequest, SourceFetchResponse,
     SourceFetchChunk, SourceFetchCredit }                 from "./mqtt-plus-msg"
@@ -143,7 +143,7 @@ export class SourceTrait<T extends APISchema = APISchema> extends ServiceTrait<T
 
             /*  determine information  */
             const requestId = request.id
-            const params    = request.params ?? []
+            const params    = alignHandlerParams(request.params ?? [], callback)
             const sender    = request.sender
             const receiver  = request.receiver